+++ packages/terra-i18n/src/I18n.js
@@ -1,7 +1,10 @@
+import { injectIntl, intlShape } from 'react-intl';
 import I18nProvider from './I18nProvider';
 import i18nLoader from './i18nLoader';
 
 module.exports = {
   I18nProvider,
   i18nLoader,
+  injectIntl,
+  intlShape,
 };